Change AddPortMapping and DeletePortMapping API
Port mapping should be a map from external port to internal but
NOT the reverse (2 different external ports can be mapped to the
same internal port). The original implementation treat it as a map
from internal port to external ports. This commit changes it to
the correct logic and updates API accordingly.

Also change a few namings to follow the Golang convention.

Signed-off-by: Yilun <zyl.skysniper@gmail.com>